+++ tests/rsocket/test_channel.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.mark.asyncio
+async def test_request_channel(pipe):
+    server, client = pipe
