+++ tests/rsocket/test_internal.py
@@ -1,0 +1,12 @@
+import asyncio
+
+import pytest
+
+
+@pytest.mark.asyncio
+async def test_reader(event_loop: asyncio.AbstractEventLoop):
+    stream = asyncio.StreamReader(loop=event_loop)
+    stream.feed_data(b'data')
+    stream.feed_eof()
+    data = await stream.read()
+    assert data == b'data'
